6_RAG_pipeline/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
psycopg2-binary
langchain
langchain-aws
langchain-core
pydantic
python-multipart
pydantic-settings
aiofiles